+++ src/utils/metadata.js
@@ -278,6 +278,7 @@ const defaultIngestMetadata = {
 			reconnect: true,
 			delay: 15,
 			staleTimeout: 30,
+			low_delay: true,
 		},
 		snapshot: {
 			enable: true,